# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APRly API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/detailed-plan:
    post:
      operationId: createDetailedPlan
      tags: [me]
      summary: Import cabinet cards and create plan leads from optimizer calculation
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportCardsInput"
      responses:
        "200":
          description: Plan leads created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateDetailedPlanResponse"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /me/partners:
    get:
      operationId: getPartners
      tags: [me]
      summary: List partners available for plan lead submission
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Partner list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerListResponse"
        "401":
          description: Unauthorized
  /me/plan-leads/{id}:
    get:
      operationId: getPlanLead
      tags: [me]
      summary: Get plan lead detail with partner and hardship portal
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Plan lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanLeadDetail"
        "401":
          description: Unauthorized
        "404":
          description: Plan lead not found

    patch:
      operationId: updatePlanLeadStatus
      tags: [me]
      summary: Update plan lead lifecycle status
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanLeadStatusInput"
      responses:
        "200":
          description: Updated plan lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanLead"
        "400":
          description: Invalid status transition
        "401":
          description: Unauthorized
        "404":
          description: Plan lead not found
  /me/plan-leads/{id}/send:
    post:
      operationId: sendPlanLead
      tags: [me]
      summary: Send plan lead to partner (simulated v1)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendPlanLeadInput"
      responses:
        "200":
          description: Plan lead sent to partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanLeadDetail"
        "400":
          description: Invalid request or status
        "401":
          description: Unauthorized
        "404":
          description: Plan lead or partner not found
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: aprly_access
  schemas:
...
    HardshipPortal:
      type: object
      required: [stage, progress, steps, etaDays]
      properties:
        stage:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        etaDays:
          type: integer
        steps:
          type: array
          items:
            $ref: "#/components/schemas/HardshipStep"
    HardshipStep:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [done, active, pending]
        description:
          type: string
        cta:
          type: string
...
    ImportCardItem:
      type: object
      required: [brand, balance, rate]
      properties:
        brand:
          type: string
          minLength: 1
        balance:
          type: number
          exclusiveMinimum: 0
        rate:
          type: number
          exclusiveMinimum: 0
          maximum: 100
        accountId:
          type: string
        source:
          type: string
          enum: [manual, plaid]
    ImportCardsInput:
      type: object
      required: [cards]
      properties:
        cards:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ImportCardItem"
...
    PlanLeadStatus:
      type: string
      enum: [recommended, in_progress, won, denied]
    LeadCardItem:
      type: object
      required:
        - id
        - brand
        - balance
        - currentApr
        - targetApr
        - estimatedAnnualSavings
      properties:
        id:
          type: integer
        brand:
          type: string
        balance:
          type: number
        currentApr:
          type: number
        targetApr:
          type: number
        estimatedAnnualSavings:
          type: number
    PlanLead:
      type: object
      description: Debt lead summary (aggregated across cards in the package)
      required:
        - id
        - brand
        - balance
        - currentApr
        - targetApr
        - estimatedAnnualSavings
        - status
        - cardCount
        - createdAt
      properties:
        id:
          type: integer
        brand:
          type: string
          description: Primary card brand or label for the lead package
        balance:
          type: number
          description: Total balance across cards
        currentApr:
          type: number
          description: Balance-weighted average APR
        targetApr:
          type: number
        estimatedAnnualSavings:
          type: number
          description: Sum of estimated savings across cards
        cardCount:
          type: integer
          minimum: 0
        status:
          $ref: "#/components/schemas/PlanLeadStatus"
        createdAt:
          type: string
          format: date-time
...
    CreateDetailedPlanResponse:
      type: object
      required: [createdCount, plans]
      properties:
        createdCount:
          type: integer
          minimum: 0
        plans:
          type: array
          items:
            $ref: "#/components/schemas/PlanLead"
    UpdatePlanLeadStatusInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/PlanLeadStatus"
    Partner:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
          description: Included in admin partner payloads
        isActive:
          type: boolean
          description: Included in admin partner payloads
...
    PartnerListResponse:
      type: object
      required: [partners]
      properties:
        partners:
          type: array
          items:
            $ref: "#/components/schemas/Partner"
    SendPlanLeadInput:
      type: object
      required: [partnerId]
      properties:
        partnerId:
          type: integer
          minimum: 1
    PlanLeadDetail:
      type: object
      description: |
        Debt lead package with card breakdown. For status `recommended`, partner fields are null.
        After POST .../send (`in_progress` or later `won`), partnerId, sentToPartnerAt,
        and partner are always set; hardshipPortal is included for in_progress and won.
      required:
        - id
        - brand
        - balance
        - currentApr
        - targetApr
        - estimatedAnnualSavings
        - status
        - cardCount
        - cards
        - createdAt
      properties:
        id:
          type: integer
        brand:
          type: string
        balance:
          type: number
        currentApr:
          type: number
        targetApr:
          type: number
        estimatedAnnualSavings:
          type: number
        cardCount:
          type: integer
          minimum: 0
        cards:
          type: array
          items:
            $ref: "#/components/schemas/LeadCardItem"
        status:
          $ref: "#/components/schemas/PlanLeadStatus"
        createdAt:
          type: string
          format: date-time
        partnerId:
          type: integer
          nullable: true
        sentToPartnerAt:
          type: string
          format: date-time
          nullable: true
        partner:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Partner"
        hardshipPortal:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/HardshipPortal"
...