# lib/api-zod/src/generated/types/planLead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APRly API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanLeadStatus } from "./planLeadStatus";

/**
 * Debt lead summary (aggregated across cards in the package)
 */
export interface PlanLead {
  id: number;
  /** Primary card brand or label for the lead package */
  brand: string;
  /** Total balance across cards */
  balance: number;
  /** Balance-weighted average APR */
  currentApr: number;
  targetApr: number;
  /** Sum of estimated savings across cards */
  estimatedAnnualSavings: number;
  /** @minimum 0 */
  cardCount: number;
  status: PlanLeadStatus;
  createdAt: Date;
}